Have Repo.__init__ convert epath to str first instead
In e133018 before this, I had noticed that Repo.__init__ implicitly
relied on the implementation detail of cygpath that it converts its
argument to str immediately. That detail was not reflected in the
parameter type annotation, which could've been broader, so I had
fixed the mypy error in Repo.__init__ by broadening cygpath's
parameter type annotation.

This undoes that change, putting the cygpath annotations back as
before, and fixes the mypy error in another way, by having
Repo.__init__ pass `str(epath)` instead of `epath` to cygpath.
The reason is the concern noted in the e133018 commit message,
that broadening the annotation made the way cygpath documents its
relationshp to Git.polish_url no longer correct, and that it is not
clear that the str(path) step in cygpath really ought to be made
more than an implementation detail (which if done may require some
documentation to be changed).

Instances of str are usually direct instances (rather than of a
subclass of str), in which case an extra call to str will (at least
in CPython, and probably in all implementations) just return the
same str object that was passed in. The performance penalty of this
extra call to str should therefore be extremely small.