Fix how Remote annotates dynamic config-backed url attribute
This fixes a mypy error by replacing an annotation on `self.url`,
which is a mypy error, by adding an annotation at the class level
for a `url` instance attribute.

This adds a corresponding brief docstring for it as well, which may
slightly improve usability, but the main impact is one less mypy
error.