Clarify CallableProgress vs. CallableRemoteProgress
This corrects an overstatement in the git.types.CallableProgress
docstring, which was introduced recently in 9e47083, and in which
I had erroneously claimed that it was the most general type of
object passed as a progress reporter for cloning.

There are some non-None non-callable types that can also be
passed and that are not encompassed by git.types.CallableProgress,
somewhat confusingly including CallableRemoteProgress (which like
RemoteProgress is not callable; rather, it wraps a callable and
forwards progress information to it).

In addition, None can be passed, and while
git.types.CallableProgress does encompass it, it is not callable.
(This is minor by comparison and I just added a brief note for it.)

This also further expands the git.types.CallableProgress docstring,
as well as the git.util.CallableRemoteProgress docstring, to
clarify the distinction between them, as well as what "Callable"
really signifies for CallableRemoteProgress (that it wraps and
forwards to a callable, rather than itself being callable).