Consolidate str and os.PathLike[str] (use GitPython's PathLike)
Where they had been introduced and written separately in 0e1df29
before this.

The main reason is that it would have to be quoted as a string
for compatibility with Python 3.8 and lower, since os.PathLike
only defined __class_getitem__ in 3.8 and later. (The definition
in git.types already does that.) In addition, the benefit of
keeping them separate was negligible, as mentioned in 0e1df29.