Spell self.Index as self.INDEX in IndexFile.diff
Because, as noted in 65863a2, it is now (slightly) preferable to
write INDEX. Note that this is solely stylstic: they are equivalent
and must remain so, so that existing code that uses GitPython and
accesses it as Index (including in overridden diff methods when
subclassing Diffable) continues to work.

This changes it in the exception message as well.