Test that redefined Diffable.Index should be compatible
This tests the most important characteristics of Diffable.Index
that must be (and are) preserved in making it an alias of the
enumeration constant INDEX instead of a nested class definition.

Adding this additional test also allows commit.INDEX to be used in
place of commit.Index in the existing test (since part of what this
tests is that they are the same object). That change is also made,
along with some very minor cleanup of immediately nearby test code.