Fix new mypy error in _read_win_env_flag
As noted in 88557bc before this, that change added a couple of new
mypy errors about unreachable code, where it is intentionally
unreachable because it is for one platform and not another.

This fixes one of them, though a fix that incorporates more of what
can now be statically known about the return value based on the
platform may be preferable.