Temporarily rename Commit_ish to Old_commit_ish
And Lit_commit_ish to Lit_old_commit_ish.

This is temporary, and only for purposes of bookkeeping while this
type is split into two types.

Specifically, Old_commit_ish will go away soon, because each
occurrence of it will be changed to one of:

- A newly redefined Commit_ish union of types representing only git
  object types that are sometimes commit-ish.

- A newly introduced union (using some name formerly not used in
  GitPython) of all four types representing git object types, the
  same types as the old Commit_ish, here renamed Old_commit_ish,
  had overbroadly covered.

- Perhaps in some cases something else.

For context, see #1858 (including comments), and comments in #1859.