Define GitObjectTypeString and update Object to use it
This is equivalent to the old Lit_commit_ish union; it is the type
of a literal string that has one of the four values that represent
actual git object types. The old Lit_commit_ish union was only used
in one place in GitPython: to annotate Object.type.

This replaces that and updates its docstring, as well as the Object
class docstring.

(One change in the Object class docstring--adding a mention of the
RootModule subclass of Submodule--is not conceptually related to
these other changes.)