Fix IndexFile.commit `parent_commits` annotation
This changes that parameter's annotation to be equivalent to the
annotation of the corresponding Commit.create_from_tree parameter.

It had been annotated as the old Commit_ish type, but it passes
that argument to Commit.create_from_tree, and it also refers to
that method in its docstring for usage. Commit.create_from_tree
annotates this argument as an optional list of Commit objects, and
if the objects obtained when it iterates them are not Commit
instances, then an exception is raised. Even if being able to pass
other commit-ish things (or references, strings, etc.) is
desirable, that does not currently work.

This also improves sorting (and grouping style) of imports in that
file (which were being changed already to no longer import the old
Commit_ish type).