Start on fixing Submodule parent_commit annotations
These used the old Commit_ish. They may not all be the same as each
other, since they perform different validation and some will look
up a commit while others do not. In particular, this is complicated
by how Submodile.__init__ documents its parent_commit parameter: it
says to see the set_parent_commit method. But __init__ sets the
_parent_commit attribute to parent_commit, while set_parent_commit
treats its commit parameter differently, calling self.repo.commit
on it to get get a Commit object from it even if it wasn't one to
begin with. See #1869 for full details.

There may be some other subtleties as well. A number of uses of the
of the old Commit_ish type appear in git.submodule.base, all
related to parent_commit. These are all remaining references to it
(identifiable due to the rename to Old_commit_ish done in 04a2753),
though. So once the changes begun here for git.submodule.* are
done, that will also have completed the broader replacement effort
begun in 7328a00. At that point cleanup can be done in git.types
(removing Old_commit_ish introduced in 04a27531, replacing
Lit_old_commit_ish with a possibly-updated and deprecated
Lit_commit_ish to avoid a breaking change, and possibly making
further refinements to additions to the newly added docstrings).