Upgrade test fixture pip in venv without upgrade_deps
Because the upgrade_deps parameter to venv.create, as well as
related functionality such as the EnvBuilder.upgrade_dependencies
method that it uses, are only available starting in Python 3.9.

This also puts the name of the VirtualEnvironment.__init__
parameter for setting up pip in the test fixture virtual
environment back from need_pip to with_pip, which may be more
intuitive.