Very slightly improve readme presentation
- Add a missing period.

- Indicate sh rather than bash as as the language for syntax
  highlighting of shell commands that don't need bash.

I had held off on making that second change in previous revisions
because it would have involved either introducing an inconsistency
or editing the section giving the deprecated signature-checking
instructions. That section was removed in 2671167 (#1823).

(This also wraps a paragraph where the immediately surrounding text
was wrapped, but that should not affect rendered text, and broader
consistency improvements to Markdown wrapping style are not done.)