Make installation instructions more consistent
If people who want to run the tests didn't install the test extra,
they can still install that extra. This simplifies the instructions
accordingly. test-requirements.txt is still mentioned near the
beginning in case people want to look at it to see dependencies.
But the changed code is the only place where the instructions had
still said to do anything with those files.

A possible disadvantage of this change is that in the rare case
that someone following those instructions to run the tests locally
didn't do an editable installation, then installing with the extra
shouldn't be done editably either. But this doesn't seem like a
likely problem, and the new text has an example command with -e to
clarify the kind of command whose effect is augmented here.

Because of that subtlety, it is not obvious that this change is
really justified for purposes of clarity. However, this also helps
prepare for a time when test-requirements.txt won't exist anymore,
as may happen when the project definition is made fully declarative
(see discussion in comments in #1716).