Colorize mypy output when run through tox
On Unix-like systems, this approach is currently only sufficient if
the TERM environment variable is set. In practice this means that
if tox is run on CI, color will not be shown. Because GitPython
does not itself use tox on CI, and other (e.g. downstream) projects
may not want color on CI or in other situations where this would be
insufficient, the further step of defining TERM as a workaround is
deliberately omitted.

See aeacb0 in #1859 for what adding TERM would look like. This does
not make any changes to how mypy runs on CI because that change is
already included there.