Let Ruff scan doc/source/conf.py
And other .py files in doc/, if there ever were any.

This may have been turned off before either for performance or out
of a concern than some flake8 plugin would catch something in the
docs themselves, but that's no longer an issue.

The doc/build directory, which should still not be scanned, is
already excluded in other ways (and does not need to be excluded
for pre-commit to improve performance, since it is untracked).