Describe Submodule.__init__ parent_commit parameter
This includes a brief description of the Submodule.__init__
parent_commit parameter in its docstring, rather than only
referring to the set_parent_commit method, whose semantics differ
due to conversation and validation, and which accepts more types
than just Commit or None.

The wording is based on wording in set_parent_commit, adjusted for
the difference in types, and set_parent_commit remains reference
for further details.

This builds on 1f03e7f (#1859) in improving the situation described
in #1869.